# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Coalition X API
  description: |    
    KPI fields support three formats (all automatically normalized): `kpi_X-X`, `X-X`, or full field names.
    See the KpiFieldMapping schema below for the complete mapping.
  version: 2.0.0
servers:
  - url: https://api.coalition-x.com
    description: Production
  - url: https://staging-api.coalition-x.com
    description: Staging
  - url: http://localhost:3000
    description: Local Development
...
paths:
...
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: User login
      description: Authenticates a user with email/password and returns OAuth tokens
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: user@example.com
                password:
                  type: string
                  format: password
                  example: securePassword123
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: User registration
      description: Registers a new user and organization, returns OAuth tokens
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
                - name
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
                name:
                  type: string
                organizationName:
                  type: string
      responses:
        '201':
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RegisterResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/auth/token:
    post:
      tags:
        - Authentication
      summary: OAuth 2.0 token endpoint
      description: |
        Exchanges credentials for access tokens. Supports multiple grant types:
        - `client_credentials`: M2M authentication via client_id + client_secret
        - `refresh_token`: Refresh an expired access token
        - `api_key`: Legacy API key exchange (backward compatibility)
        
        Also accepts `X-API-Key` header for backward compatibility.
      operationId: getToken
      parameters:
        - name: X-API-Key
          in: header
          description: API key for legacy authentication (alternative to grant_type=api_key)
          schema:
            type: string
          required: false
      requestBody:
        required: false
        content:
          application/json:
            schema:
              oneOf:
                - $ref: '#/components/schemas/ClientCredentialsRequest'
                - $ref: '#/components/schemas/RefreshTokenRequest'
                - $ref: '#/components/schemas/ApiKeyRequest'
          application/x-www-form-urlencoded:
            schema:
              oneOf:
                - $ref: '#/components/schemas/ClientCredentialsRequest'
                - $ref: '#/components/schemas/RefreshTokenRequest'
                - $ref: '#/components/schemas/ApiKeyRequest'
      responses:
        '200':
          description: Token issued successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/partner-requests:
    post:
      tags:
        - Partner Requests
      summary: Submit partner application
      description: Submits a new partner application (public endpoint, no auth required)
      operationId: submitPartnerRequest
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePartnerRequest'
      responses:
        '201':
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PartnerRequestCreated'
        '400':
          $ref: '#/components/responses/BadRequest'
    get:
      tags:
        - Partner Requests
      summary: List partner applications
      description: Lists partner applications (admin only)
      operationId: listPartnerRequests
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, APPROVED, REJECTED]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of applications
          content:
            application/json:
              schema:
                type: object
                properties:
                  applications:
                    type: array
                    items:
                      $ref: '#/components/schemas/PartnerRequest'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
  /api/admin/organizations:
    get:
      tags:
        - Admin
      summary: List organizations
      description: Lists all ACCREDITED_PARTNER organizations (admin only)
      operationId: listOrganizations
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [PENDING, ACTIVE, SUSPENDED, INACTIVE]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: List of organizations
          content:
            application/json:
              schema:
                type: object
                properties:
                  organizations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Organization'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags:
        - Admin
      summary: Create organization
      description: Creates a new organization (admin only)
      operationId: createOrganization
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateOrganizationRequest'
      responses:
        '201':
          description: Organization created
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  organization:
                    $ref: '#/components/schemas/Organization'
                  credentials:
                    type: object
                    description: Only present if type is ACCREDITED_PARTNER
                    properties:
                      client_id:
                        type: string
                      client_secret:
                        type: string
                  warning:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OAuth 2.0 JWT access token
    CookieAuth:
      type: apiKey
      in: cookie
      name: session
      description: Session cookie for web UI
  schemas:
...
    ClientCredentialsRequest:
      type: object
      required:
        - grant_type
        - client_id
        - client_secret
      properties:
        grant_type:
          type: string
          enum: [client_credentials]
        client_id:
          type: string
        client_secret:
          type: string
        scope:
          type: string
    RefreshTokenRequest:
      type: object
      required:
        - grant_type
        - refresh_token
      properties:
        grant_type:
          type: string
          enum: [refresh_token]
        refresh_token:
          type: string
    ApiKeyRequest:
      type: object
      required:
        - grant_type
        - api_key
      properties:
        grant_type:
          type: string
          enum: [api_key]
        api_key:
          type: string
    TokenResponse:
      type: object
      properties:
        access_token:
          type: string
        refresh_token:
          type: string
        token_type:
          type: string
          enum: [Bearer]
        expires_in:
          type: integer
          description: Token lifetime in seconds
        expires_at:
          type: string
          format: date-time
        scope:
          type: string
    LoginResponse:
      allOf:
        - $ref: '#/components/schemas/TokenResponse'
        - type: object
          properties:
            user:
              type: object
              properties:
                id:
                  type: string
                email:
                  type: string
                name:
                  type: string
                  nullable: true
                role:
                  type: string
            organization:
              type: object
              nullable: true
              properties:
                id:
                  type: string
                name:
                  type: string
                type:
                  type: string
    RegisterResponse:
      allOf:
        - $ref: '#/components/schemas/TokenResponse'
        - type: object
          properties:
            user:
              type: object
              properties:
                id:
                  type: string
                email:
                  type: string
                name:
                  type: string
                  nullable: true
            organization:
              type: object
              properties:
                id:
                  type: string
                name:
                  type: string
...
    CreatePartnerRequest:
      type: object
      required:
        - companyName
        - contactName
        - contactEmail
        - integrationType
        - useCase
      properties:
        companyName:
          type: string
        contactName:
          type: string
        contactEmail:
          type: string
          format: email
        website:
          type: string
        integrationType:
          type: string
          enum: [API, MANUAL, HYBRID]
        estimatedVolume:
          type: integer
        useCase:
          type: string
    PartnerRequest:
      type: object
      properties:
        id:
          type: string
        companyName:
          type: string
        contactName:
          type: string
        contactEmail:
          type: string
        website:
          type: string
          nullable: true
        integrationType:
          type: string
        estimatedVolume:
          type: integer
          nullable: true
        useCase:
          type: string
        status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
        reviewNotes:
          type: string
          nullable: true
        rejectionReason:
          type: string
          nullable: true
        organizationId:
          type: string
          nullable: true
        requestedAt:
          type: string
          format: date-time
        reviewedAt:
          type: string
          format: date-time
          nullable: true
        approvedAt:
          type: string
          format: date-time
          nullable: true
    PartnerRequestCreated:
      type: object
      properties:
        applicationId:
          type: string
        status:
          type: string
          enum: [PENDING]
        message:
          type: string
...
    CreateOrganizationRequest:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
        type:
          type: string
          enum: [ACCREDITED_PARTNER, CLIENT]
        contactEmail:
          type: string
          format: email
        contactName:
          type: string
        description:
          type: string
...
    Organization:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        name:
          type: string
        type:
          type: string
          enum: [ACCREDITED_PARTNER, CLIENT, ADMIN]
        status:
          type: string
          enum: [PENDING, ACTIVE, SUSPENDED, INACTIVE]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        userCount:
          type: integer
        assetCount:
          type: integer
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: string
        details:
          type: object

    # KPI Field Mapping
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized - Invalid or missing token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Forbidden - Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...